rename `useLocation` to `useBrowserLocation`
