+++ packages/wouter/test/use-browser-location.test.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { it, expect, describe, beforeEach } from "vitest";
 import { renderHook, act, waitFor } from "@testing-library/react";
-import { useLocation, navigate, useSearch } from "wouter/use-location";
+import { useBrowserLocation as useLocation, navigate, useSearch } from "wouter/use-browser-location";
 
 it("returns a pair [value, update]", () => {
   const { result, unmount } = renderHook(() => useLocation());
